requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-flask>=1.3.0
factory-boy>=3.3.0
responses>=0.23.0
mongomock>=4.1.0
fakeredis>=2.18.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0
isort>=5.12.0

# Development Tools
ipython>=8.15.0
ipdb>=0.13.0
python-dotenv>=1.0.0